added 2 methods:
1) updateMessageFromPutRequest
recommended method. Just update the message of the day by put request.

2) updateMessageFromGetRequest
Not recommended, but might be used in case if the user does not have any tools like a curl or telnet
but have to be able to update message.
In this case, the message might be updated using the link in any browser
for example "/updateMessage?newMessageOfTheDay=Hello World!"